+++ release-notes.md
@@ -1,3 +1,12 @@
+Version 1.1.0
+--------------
+* Added support for GLTF extensions as user-added data on the
+    Spatia.  Includes support for all JSON primitives as well
+    as Lists and Maps.
+* Added a 'u' option to the -probe to show the user-added data
+    of a converted model. 
+
+
 Version 1.0.0
 --------------
 * Initial release, includes: